rename the member list components
