+++ lib/widgets/pages/member_list/member_list_page.dart
@@ -15,9 +15,9 @@ import 'package:weforza/widgets/common/rider_search_filter_empty.dart';
 import 'package:weforza/widgets/pages/add_member/add_member_page.dart';
 import 'package:weforza/widgets/pages/export_members_page.dart';
 import 'package:weforza/widgets/pages/import_members_page.dart';
-import 'package:weforza/widgets/pages/memberDetails/member_details_page.dart';
-import 'package:weforza/widgets/pages/memberList/memberListEmpty.dart';
-import 'package:weforza/widgets/pages/memberList/memberListItem.dart';
+import 'package:weforza/widgets/pages/member_details/member_details_page.dart';
+import 'package:weforza/widgets/pages/member_list/member_list_empty.dart';
+import 'package:weforza/widgets/pages/member_list/member_list_item.dart';
 import 'package:weforza/widgets/platform/cupertinoIconButton.dart';
 import 'package:weforza/widgets/platform/platformAwareLoadingIndicator.dart';
 import 'package:weforza/widgets/platform/platformAwareWidget.dart';
